+++ packages/spright-components/src/chat/input/index.ts
@@ -21,11 +21,20 @@ export class ChatInput extends FoundationElement {
     @attr({ attribute: 'send-button-label' })
     public sendButtonLabel?: string;
 
+    @attr({ attribute: 'stop-button-label' })
+    public stopButtonLabel?: string;
+
     @attr
     public value = '';
 
     @attr({ attribute: 'tabindex', converter: nullableNumberConverter })
     public override tabIndex!: number;
+
+    @attr({ attribute: 'maxlength', converter: nullableNumberConverter })
+    public maxLength?: number = -1;
+
+    @attr({ attribute: 'processing', mode: 'boolean' })
+    public processing = false;
 
     /**
      * @internal
@@ -44,6 +53,9 @@ export class ChatInput extends FoundationElement {
      */
     public textAreaKeydownHandler(e: KeyboardEvent): boolean {
         if (e.key === keyEnter && !e.shiftKey) {
+            if (this.processing) {
+                return false;
+            }
             this.sendButtonClickHandler();
             return false;
         }
@@ -89,6 +101,17 @@ export class ChatInput extends FoundationElement {
         };
         this.resetInput();
         this.$emit('send', eventDetail);
+    }
+
+    /**
+     * @internal
+     */
+    public stopButtonClickHandler(): void {
+        if (!this.processing) {
+            return;
+        }
+        this.$emit('stop');
+        this.textArea?.blur();
     }
 
     private shouldDisableSendButton(): boolean {